Get reaction chain for specific mass to charge ratio

Get reaction chain for specific mass to charge ratio

getchain(
  list,
  diff,
  mass,
  digits = 2,
  accuracy = 4,
  rtcutoff = 10,
  corcutoff = 0.6,
  ppm = 25
)

Arguments

list

a list with mzrt profile

diff

paired mass distance(s) of interests

mass

a specific mass for known compound or a vector of masses. You could also input formula for certain compounds

digits

mass or mass to charge ratio accuracy for pmd, default 2

accuracy

measured mass or mass to charge ratio in digits, default 4

rtcutoff

cutoff of the distances in retention time hierarchical clustering analysis, default 10

corcutoff

cutoff of the correlation coefficient, default NULL

ppm

all the peaks within this mass accuracy as seed mass or formula

Value

a list with mzrt profile and reaction chain dataframe

Examples

data(spmeinvivo)
# check metabolites of C18H39NO
pmd <- getchain(spmeinvivo,diff = c(2.02,14.02,15.99),mass = 286.3101)