Get quantitative paired peaks list for specific reaction/pmd

Get quantitative paired peaks list for specific reaction/pmd

getreact(list, pmd, rtcutoff = 10, digits = 2, accuracy = 4,
  ratiocv = 30, outlier = F, ...)

Arguments

list
a list with mzrt profile and data

pmd
a specific paired mass distances

rtcutoff
cutoff of the distances in retention time hierarchical clustering analysis, default 10

digits
mass or mass to charge ratio accuracy for pmd, default 2

accuracy
measured mass or mass to charge ratio in digits, default 4

ratiocv
ratio cv cutoff for quantitative paired peaks, default 30

outlier
logical, if true, outliar of ratio will be removed, default False.

...
other parameters for getpmd

Value

list with quantitative paired peaks.

See also

Examples

data(spmeinvivo)
pmd <- getreact(spmeinvivo,pmd=15.99)