GlobalStd algorithm with structure/reaction directed analysis

GlobalStd algorithm with structure/reaction directed analysis

globalstd(
  list,
  rtcutoff = 10,
  ng = 10,
  corcutoff = NULL,
  digits = 2,
  accuracy = 4,
  freqcutoff = NULL
)

Arguments

list
a peaks list with mass to charge, retention time and intensity data

rtcutoff
cutoff of the distances in cluster, default 10

ng
cutoff of global PMD's retention time group numbers, default 10. If ng = NULL, 20 percent of RT cluster will be used as ng.

corcutoff
cutoff of the correlation coefficient, default NULL

digits
mass or mass to charge ratio accuracy for pmd, default 2

accuracy
measured mass or mass to charge ratio in digits, default 4

freqcutoff
pmd freqency cutoff for structures or reactions, default NULL. This cutoff will be found by PMD network analysis when it is NULL.

Value

list with GlobalStd algorithm processed data.

See also

Examples

data(spmeinvivo)
re <- globalstd(spmeinvivo)

#> 75 retention time cluster found.

#> 380 paired masses found 

#> 9 unique within RT clusters high frequency PMD(s) used for further investigation.

#> 719 isotopologue(s) related paired mass found.

#> 492 multi-charger(s) related paired mass found.

#> 8 retention group(s) have single peaks. 14 23 32 33 54 55 56 75

#> 11 group(s) with multiple peaks while no isotope/paired relationship 4 5 7 8 11 41 42 49 68 72 73

#> 9 group(s) with multiple peaks with isotope without paired relationship 2 9 22 26 52 62 64 66 70

#> 4 group(s) with paired relationship without isotope 1 10 15 18

#> 43 group(s) with paired relationship and isotope 3 6 12 13 16 17 19 20 21 24 25 27 28 29 30 31 34 35 36 37 38 39 40 43 44 45 46 47 48 50 51 53 57 58 59 60 61 63 65 67 69 71 74

#> 297 std mass found.

#> PMD frequency cutoff is 9 by PMD network analysis with 23 clusters.

#> 21 groups were found as high frequency PMD group.

#> 0 was found as high frequency PMD. 
#> 1.98 was found as high frequency PMD. 
#> 2.02 was found as high frequency PMD. 
#> 12 was found as high frequency PMD. 
#> 13.98 was found as high frequency PMD. 
#> 14.02 was found as high frequency PMD. 
#> 14.05 was found as high frequency PMD. 
#> 15.99 was found as high frequency PMD. 
#> 16.03 was found as high frequency PMD. 
#> 28.03 was found as high frequency PMD. 
#> 30.05 was found as high frequency PMD. 
#> 42.05 was found as high frequency PMD. 
#> 49.02 was found as high frequency PMD. 
#> 58.04 was found as high frequency PMD. 
#> 66.05 was found as high frequency PMD. 
#> 68.06 was found as high frequency PMD. 
#> 74.02 was found as high frequency PMD. 
#> 82.08 was found as high frequency PMD. 
#> 88.05 was found as high frequency PMD. 
#> 116.08 was found as high frequency PMD. 
#> 126.14 was found as high frequency PMD.